ManageAddView base class
