+++ bika/lims/browser/analysisrequest/add2.py
@@ -8,7 +8,6 @@
 from datetime import datetime
 
 import magnitude
-from BTrees.OOBTree import OOBTree
 from DateTime import DateTime
 from Products.CMFPlone.utils import _createObjectByType
 from Products.CMFPlone.utils import safe_unicode
@@ -16,7 +15,6 @@ from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from plone import protect
 from plone.memoize.volatile import cache
-from zope.annotation.interfaces import IAnnotations
 from zope.i18n.locales import locales
 from zope.interface import implements
 from zope.publisher.interfaces import IPublishTraverse
@@ -24,13 +22,11 @@ from zope.publisher.interfaces import IPublishTraverse
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
+from bika.lims.browser.base_manage_add_view import BaseManageAddView
 from bika.lims.utils import cache_key
 from bika.lims.utils import returns_json
 from bika.lims.utils import tmpID
 from bika.lims.utils.analysisrequest import create_analysisrequest as crar
-
-AR_CONFIGURATION_STORAGE = "bika.lims.browser.analysisrequest.manage.add"
-SKIP_FIELD_ON_COPY = ["Sample"]
 
 
 def mg(value):
@@ -641,125 +637,26 @@ class AnalysisRequestAddView(BrowserView):
         return False
 
 
-class AnalysisRequestManageView(BrowserView):
+class AnalysisRequestManageView(BaseManageAddView):
     """AR Manage View
     """
     template = ViewPageTemplateFile("templates/ar_add_manage.pt")
 
     def __init__(self, context, request):
-        self.context = context
-        self.request = request
-        self.tmp_ar = None
+        BaseManageAddView.__init__(self, context, request)
+        self.CONFIGURATION_STORAGE = \
+            "bika.lims.browser.analysisrequest.manage.add"
+        self.SKIP_FIELD_ON_COPY = ["Sample"]
 
     def __call__(self):
-        protect.CheckAuthenticator(self.request.form)
-        form = self.request.form
-        if form.get("submitted", False) and form.get("save", False):
-            order = form.get("order")
-            self.set_field_order(order)
-            visibility = form.get("visibility")
-            self.set_field_visibility(visibility)
-        if form.get("submitted", False) and form.get("reset", False):
-            self.flush()
+        BaseManageAddView.__call__(self)
         return self.template()
 
-    def get_ar(self):
-        if not self.tmp_ar:
-            self.tmp_ar = self.context.restrictedTraverse("portal_factory/AnalysisRequest/Request new analyses")
-        return self.tmp_ar
-
-    def get_annotation(self):
-        bika_setup = api.get_bika_setup()
-        return IAnnotations(bika_setup)
-
-    @property
-    def storage(self):
-        annotation = self.get_annotation()
-        if annotation.get(AR_CONFIGURATION_STORAGE) is None:
-            annotation[AR_CONFIGURATION_STORAGE] = OOBTree()
-        return annotation[AR_CONFIGURATION_STORAGE]
-
-    def flush(self):
-        annotation = self.get_annotation()
-        if annotation.get(AR_CONFIGURATION_STORAGE) is not None:
-            del annotation[AR_CONFIGURATION_STORAGE]
-
-    def set_field_order(self, order):
-        self.storage.update({"order": order})
-
-    def get_field_order(self):
-        order = self.storage.get("order")
-        if order is None:
-            return map(lambda f: f.getName(), self.get_fields())
-        return order
-
-    def set_field_visibility(self, visibility):
-        self.storage.update({"visibility": visibility})
-
-    def get_field_visibility(self):
-        return self.storage.get("visibility")
-
-    def is_field_visible(self, field):
-        if field.required:
-            return True
-        visibility = self.get_field_visibility()
-        if visibility is None:
-            return True
-        return visibility.get(field.getName(), True)
-
-    def get_field(self, name):
-        """Get AR field by name
-        """
-        ar = self.get_ar()
-        return ar.getField(name)
-
-    def get_fields(self):
-        """Return all AR fields
-        """
-        ar = self.get_ar()
-        return ar.Schema().fields()
-
-    def get_sorted_fields(self):
-        """Return the sorted fields
-        """
-        inf = float("inf")
-        order = self.get_field_order()
-
-        def field_cmp(field1, field2):
-            _n1 = field1.getName()
-            _n2 = field2.getName()
-            _i1 = _n1 in order and order.index(_n1) + 1 or inf
-            _i2 = _n2 in order and order.index(_n2) + 1 or inf
-            return cmp(_i1, _i2)
-
-        return sorted(self.get_fields(), cmp=field_cmp)
-
-    def get_fields_with_visibility(self, visibility="edit", mode="add"):
-        """Return the fields with visibility
-        """
-        fields = self.get_sorted_fields()
-
-        out = []
-
-        for field in fields:
-            v = field.widget.isVisible(
-                self.context, mode, default='invisible', field=field)
-
-            if self.is_field_visible(field) is False:
-                v = "hidden"
-
-            visibility_guard = True
-            # visibility_guard is a widget field defined in the schema in order
-            # to know the visibility of the widget when the field is related to
-            # a dynamically changing content such as workflows. For instance
-            # those fields related to the workflow will be displayed only if
-            # the workflow is enabled, otherwise they should not be shown.
-            if 'visibility_guard' in dir(field.widget):
-                visibility_guard = eval(field.widget.visibility_guard)
-            if v == visibility and visibility_guard:
-                out.append(field)
-
-        return out
+    def get_obj(self):
+        if not self.tmp_obj:
+            self.tmp_obj = self.context.restrictedTraverse(
+                "portal_factory/AnalysisRequest/Request new analyses")
+        return self.tmp_obj
 
 
 class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
