+++ bika/lims/browser/base_manage_add_view.py
@@ -1,0 +1,141 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of SENAITE.CORE
+#
+# Copyright 2018 by it's authors.
+# Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
+
+from BTrees.OOBTree import OOBTree
+from Products.Five.browser import BrowserView
+from plone import protect
+from zope.annotation.interfaces import IAnnotations
+
+from bika.lims import api
+
+
+class BaseManageAddView(BrowserView):
+    """
+    Base view to create a visibility and order manager for an Add View.
+    """
+    template = None
+
+    def __init__(self, context, request):
+        BrowserView.__init__(self, context, request)
+        self.context = context
+        self.request = request
+        self.tmp_obj = None
+        self.CONFIGURATION_STORAGE = None
+        self.SKIP_FIELD_ON_COPY = []
+
+    def __call__(self):
+        protect.CheckAuthenticator(self.request.form)
+        form = self.request.form
+        if form.get("submitted", False) and form.get("save", False):
+            order = form.get("order")
+            self.set_field_order(order)
+            visibility = form.get("visibility")
+            self.set_field_visibility(visibility)
+        if form.get("submitted", False) and form.get("reset", False):
+            self.flush()
+        return self.template()
+
+    def get_obj(self):
+        """
+        This method should return a temporary object. This object should be
+        of the same type as the one we are creating.
+        :return: ATObjectType
+        """
+        raise NotImplementedError("get_obj is not implemented.")
+
+    def get_annotation(self):
+        bika_setup = api.get_bika_setup()
+        return IAnnotations(bika_setup)
+
+    @property
+    def storage(self):
+        annotation = self.get_annotation()
+        if annotation.get(self.CONFIGURATION_STORAGE) is None:
+            annotation[self.CONFIGURATION_STORAGE] = OOBTree()
+        return annotation[self.CONFIGURATION_STORAGE]
+
+    def flush(self):
+        annotation = self.get_annotation()
+        if annotation.get(self.CONFIGURATION_STORAGE) is not None:
+            del annotation[self.CONFIGURATION_STORAGE]
+
+    def set_field_order(self, order):
+        self.storage.update({"order": order})
+
+    def get_field_order(self):
+        order = self.storage.get("order")
+        if order is None:
+            return map(lambda f: f.getName(), self.get_fields())
+        return order
+
+    def set_field_visibility(self, visibility):
+        self.storage.update({"visibility": visibility})
+
+    def get_field_visibility(self):
+        return self.storage.get("visibility")
+
+    def is_field_visible(self, field):
+        if field.required:
+            return True
+        visibility = self.get_field_visibility()
+        if visibility is None:
+            return True
+        return visibility.get(field.getName(), True)
+
+    def get_field(self, name):
+        """Get a field of the object by name
+        """
+        obj = self.get_obj()
+        return obj.getField(name)
+
+    def get_fields(self):
+        """Return all Object fields
+        """
+        obj = self.get_obj()
+        return obj.Schema().fields()
+
+    def get_sorted_fields(self):
+        """Return the sorted fields
+        """
+        inf = float("inf")
+        order = self.get_field_order()
+
+        def field_cmp(field1, field2):
+            _n1 = field1.getName()
+            _n2 = field2.getName()
+            _i1 = _n1 in order and order.index(_n1) + 1 or inf
+            _i2 = _n2 in order and order.index(_n2) + 1 or inf
+            return cmp(_i1, _i2)
+
+        return sorted(self.get_fields(), cmp=field_cmp)
+
+    def get_fields_with_visibility(self, visibility="edit", mode="add"):
+        """Return the fields with visibility
+        """
+        fields = self.get_sorted_fields()
+
+        out = []
+
+        for field in fields:
+            v = field.widget.isVisible(
+                self.context, mode, default='invisible', field=field)
+
+            if self.is_field_visible(field) is False:
+                v = "hidden"
+
+            visibility_guard = True
+            # visibility_guard is a widget field defined in the schema in order
+            # to know the visibility of the widget when the field is related to
+            # a dynamically changing content such as workflows. For instance
+            # those fields related to the workflow will be displayed only if
+            # the workflow is enabled, otherwise they should not be shown.
+            if 'visibility_guard' in dir(field.widget):
+                visibility_guard = eval(field.widget.visibility_guard)
+            if v == visibility and visibility_guard:
+                out.append(field)
+
+        return out
